better version of cat and cache paths
