+++ model_catalogs/model_catalogs/__init__.py
@@ -2,6 +2,7 @@
 Set up for using package.
 """
 
+import importlib
 import shutil
 import pandas as pd
 
@@ -11,7 +12,6 @@ from appdirs import AppDirs
 from pkg_resources import DistributionNotFound, get_distribution
 
 from .model_catalogs import (  # noqa
-    calculate_boundaries,
     find_availability,
     make_catalog,
     select_date_range,
@@ -21,6 +21,7 @@ from .model_catalogs import (  # noqa
 from .utils import (  # noqa
     agg_for_date,
     astype,
+    calculate_boundaries,
     find_bbox,
     find_catrefs,
     find_filelocs,
@@ -36,73 +37,78 @@ except DistributionNotFound:
     # package is not installed
     __version__ = "unknown"
 
+# set up known locations for catalogs. Can be overwritten HOW
+# this is where the original model catalog files and previously-calculated
+# model boundaries can be found, which are hard-wired in the repo
+# CAT_PATH = importlib.resources.path('model_catalogs', 'catalogs')
+# CAT_PATH = Path(__file__).parent / "catalogs"
+with importlib.resources.path('model_catalogs', 'catalogs') as pth:
+    CAT_PATH = pth
+CAT_PATH_ORIG = CAT_PATH / "orig"
+CAT_PATH_BOUNDARIES = CAT_PATH / "boundaries"
+CAT_PATH_TRANSFORM = CAT_PATH / "transform.yaml"
+
 # set up cache directories for package to use
 # user application cache directory, appropriate to each OS
 dirs = AppDirs("model_catalogs", "NOAA-ORR-ERD")
 cache_dir = Path(dirs.user_cache_dir)
 
-# set up known locations for catalogs. Can be overwritten HOW
-# By default, put catalog directory in home directory
-CAT_PATH = cache_dir / "catalogs"
-# CAT_PATH = Path.home() / "catalogs"
-# UPDATE THESE
-SOURCE_CATALOG_NAME = "source_catalog.yaml"
-CAT_PATH_ORIG = CAT_PATH / "orig"
-CAT_PATH_BOUNDARY = CAT_PATH / "boundaries"
-CAT_PATH_COMPILED = CAT_PATH / "compiled"
-CAT_PATH_AVAILABILITY = CAT_PATH / "availability"
-CAT_PATH_FILE_LOCS = CAT_PATH / "file_locs"
-SOURCE_TRANSFORM = CAT_PATH / "transform.yaml"
+# This is where files are saved to refer back to for saving time
+CACHE_PATH = cache_dir / "catalogs"
+# SOURCE_CATALOG_NAME = "source_catalog.yaml"
+CACHE_PATH_COMPILED = CACHE_PATH / "compiled"
+CACHE_PATH_AVAILABILITY = CACHE_PATH / "availability"
+CACHE_PATH_FILE_LOCS = CACHE_PATH / "file_locs"
 
 # make directories
 # CAT_PATH.mkdir(parents=True, exist_ok=True)
-CAT_PATH_ORIG.mkdir(parents=True, exist_ok=True)
-CAT_PATH_BOUNDARY.mkdir(parents=True, exist_ok=True)
-CAT_PATH_AVAILABILITY.mkdir(parents=True, exist_ok=True)
-CAT_PATH_COMPILED.mkdir(parents=True, exist_ok=True)
-CAT_PATH_FILE_LOCS.mkdir(parents=True, exist_ok=True)
+# CAT_PATH_ORIG.mkdir(parents=True, exist_ok=True)
+# CAT_PATH_BOUNDARIES.mkdir(parents=True, exist_ok=True)
+CACHE_PATH_COMPILED.mkdir(parents=True, exist_ok=True)
+CACHE_PATH_AVAILABILITY.mkdir(parents=True, exist_ok=True)
+CACHE_PATH_FILE_LOCS.mkdir(parents=True, exist_ok=True)
 
-# Move "orig" catalog files to catalog dir
-PKG_CAT_PATH_ORIG = Path(__path__[0]) / "catalogs" / "orig"
-[
-    shutil.copy(fname, CAT_PATH_ORIG)
-    for fname in PKG_CAT_PATH_ORIG.glob("*.yaml")
-]
-PKG_CAT_PATH_BOUNDARY = Path(__path__[0]) / "catalogs" / "boundaries"
-[
-    shutil.copy(fname, CAT_PATH_BOUNDARY)
-    for fname in PKG_CAT_PATH_BOUNDARY.glob("*.yaml")
-]
+# # Move "orig" catalog files to catalog dir
+# PKG_CAT_PATH_ORIG = Path(__path__[0]) / "catalogs" / "orig"
+# [
+#     shutil.copy(fname, CAT_PATH_ORIG)
+#     for fname in PKG_CAT_PATH_ORIG.glob("*.yaml")
+# ]
+# PKG_CAT_PATH_BOUNDARY = Path(__path__[0]) / "catalogs" / "boundaries"
+# [
+#     shutil.copy(fname, CAT_PATH_BOUNDARIES)
+#     for fname in PKG_CAT_PATH_BOUNDARY.glob("*.yaml")
+# ]
 
-# Move "transform.yaml" to catalog dir
-SOURCE_TRANSFORM_REPO = Path(__path__[0]) / "catalogs" / "transform.yaml"
-shutil.copy(SOURCE_TRANSFORM_REPO, SOURCE_TRANSFORM)
+# # Move "transform.yaml" to catalog dir
+# SOURCE_TRANSFORM_REPO = Path(__path__[0]) / "catalogs" / "transform.yaml"
+# shutil.copy(SOURCE_TRANSFORM_REPO, SOURCE_TRANSFORM)
 
 
 def FILE_PATH_COMPILED(modelyaml):
     """Return filename for model boundaries information."""
-    return CAT_PATH_COMPILED / modelyaml
+    return CACHE_PATH_COMPILED / modelyaml
 
 
 # availability file names
 def FILE_PATH_START(model, timing):
     """Return filename for model/timing start time."""
-    return CAT_PATH_AVAILABILITY / f"{model}_{timing}_start_datetime.yaml"
+    return CACHE_PATH_AVAILABILITY / f"{model}_{timing}_start_datetime.yaml"
 
 
 def FILE_PATH_END(model, timing):
     """Return filename for model/timing end time."""
-    return CAT_PATH_AVAILABILITY / f"{model}_{timing}_end_datetime.yaml"
+    return CACHE_PATH_AVAILABILITY / f"{model}_{timing}_end_datetime.yaml"
 
 
-def FILE_PATH_BOUNDARY(modelyaml):
+def FILE_PATH_BOUNDARIES(modelyaml):
     """Return filename for model boundaries information."""
-    return CAT_PATH_BOUNDARY / modelyaml
+    return CAT_PATH_BOUNDARIES / modelyaml
 
 
 def FILE_PATH_CATREFS(model, timing):
     """Return filename for model/timing start time."""
-    return CAT_PATH_AVAILABILITY / f"{model}_{timing}_catrefs.yaml"
+    return CACHE_PATH_AVAILABILITY / f"{model}_{timing}_catrefs.yaml"
 
 
 def FILE_PATH_AGG_FILE_LOCS(model, timing, date, is_fore):
@@ -111,7 +117,7 @@ def FILE_PATH_AGG_FILE_LOCS(model, timing, date, is_fore):
     Date included to day."""
     date = astype(date, pd.Timestamp)
     name = f"{model}_{timing}_{date.isoformat()[:10]}_is-forecast_{is_fore}.yaml"
-    return CAT_PATH_FILE_LOCS / name
+    return CACHE_PATH_FILE_LOCS / name
 
 
 # Fresh parameters: how long until model output avialability will be refreshed
